+++ sdk/src/main/java/com/vk/api/sdk/objects/groups/GroupsArray.java
@@ -21,7 +21,7 @@ public class GroupsArray implements Validable {
 
     @SerializedName("items")
     @Required
-    private List<Integer> items;
+    private List<Long> items;
 
     public Integer getCount() {
         return count;
@@ -32,11 +32,11 @@ public class GroupsArray implements Validable {
         return this;
     }
 
-    public List<Integer> getItems() {
+    public List<Long> getItems() {
         return items;
     }
 
-    public GroupsArray setItems(List<Integer> items) {
+    public GroupsArray setItems(List<Long> items) {
         this.items = items;
         return this;
     }
