+++ sdk/src/main/java/com/vk/api/sdk/objects/likes/responses/GetListResponse.java
@@ -21,7 +21,7 @@ public class GetListResponse implements Validable {
 
     @SerializedName("items")
     @Required
-    private List<Integer> items;
+    private List<Long> items;
 
     public Integer getCount() {
         return count;
@@ -32,11 +32,11 @@ public class GetListResponse implements Validable {
         return this;
     }
 
-    public List<Integer> getItems() {
+    public List<Long> getItems() {
         return items;
     }
 
-    public GetListResponse setItems(List<Integer> items) {
+    public GetListResponse setItems(List<Long> items) {
         this.items = items;
         return this;
     }
