+++ sdk/src/main/java/com/vk/api/sdk/objects/users/UsersArray.java
@@ -21,7 +21,7 @@ public class UsersArray implements Validable {
 
     @SerializedName("items")
     @Required
-    private List<Integer> items;
+    private List<Long> items;
 
     public Integer getCount() {
         return count;
@@ -32,11 +32,11 @@ public class UsersArray implements Validable {
         return this;
     }
 
-    public List<Integer> getItems() {
+    public List<Long> getItems() {
         return items;
     }
 
-    public UsersArray setItems(List<Integer> items) {
+    public UsersArray setItems(List<Long> items) {
         this.items = items;
         return this;
     }
